grt: inline HeatMapHandleAdapter into MakeGlobalRouter.cpp
The clang-tidy CI action lints every changed file including headers,
but HeatMapHandleAdapter.h had no compile_commands.json entry (only
TUs do), so clang-tidy ran with no -I flags and failed to resolve
gui/heatMap.h. The adapter has a single user; move the class into
MakeGlobalRouter.cpp and delete the header.
