+++ src/grt/src/MakeGlobalRouter.cpp
@@ -4,10 +4,12 @@
 #include "grt/MakeGlobalRouter.h"
 
 #include <memory>
+#include <utility>
 
+#include "AbstractRoutingCongestionDataSource.h"
 #include "FastRoute.h"
-#include "HeatMapHandleAdapter.h"
 #include "grt/GlobalRouter.h"
+#include "gui/heatMap.h"
 #include "heatMap.h"
 #include "heatMapRudy.h"
 #include "tcl.h"
@@ -21,6 +23,27 @@ namespace grt {
 
 // Tcl files encoded into strings.
 extern const char* grt_tcl_inits[];
+
+// Adapts a gui::HeatMapSourceHandle to the gui-free abstract interface
+// consumed by GlobalRouter. Defined in this TU so :grt does not see gui.
+class HeatMapHandleAdapter : public AbstractRoutingCongestionDataSource
+{
+ public:
+  explicit HeatMapHandleAdapter(gui::HeatMapSourceHandle handle)
+      : handle_(std::move(handle))
+  {
+  }
+
+  void invalidate() override
+  {
+    if (handle_) {
+      handle_->invalidateInstances();
+    }
+  }
+
+ private:
+  gui::HeatMapSourceHandle handle_;
+};
 
 void initGui(grt::GlobalRouter* grt, odb::dbDatabase* db, utl::Logger* logger)
 {
